fix(search): a cancelled index lap is not an empty one
Eighteen lint findings — the brief said nine, and the gap is
golangci-lint's `max-same-issues` collapsing identical messages again.

THE REAL ONE IS A nilerr. Indexer.staleIn checked ctx.Err() at the
bottom of its lap and answered (nil, nil). That is the SAME ANSWER as
"this source has nothing stale in it", and it is exactly the licence
Indexer.Sweep hands its caller: `false` means a whole lap found nothing,
go to sleep. So a walk that was STOPPED part way told a fixed-point
driver — indexAll in the tests, Run in production — that the index had
caught up with a corpus it never finished reading. It returns the
context error now, and the two honest exits from that loop are "it
wrapped" and "it found work".

`normalise` in the fixture is deleted rather than wired up, and the
reasoning is the fixture's own: it never needs a unit-length vector.
Quantize is a SIGN code and therefore scale-invariant, and the exact
similarity is computed from the coefficients divided by the stored norms
— so normalising a composed vector could not change one bit of what the
deterministic gate measures. `norm` itself stays; it has three callers.

`int64(at.Packed())` twice in apply.go, where statelog.Position.Packed()
already returns int64. Not a behaviour bug, but it reads as a width
claim about a value that has none, and internal/tracker/apply.go writes
`at.Packed()` bare.

SearchQuery and SearchHit are now LexicalQuery and LexicalHit. Renaming
was the right half of revive's stutter finding rather than documenting
it: the package is `search`, so `search.SearchQuery` said the word
twice, and `Lexical` is what these two actually are — the semantic half
has its own shapes. Swept repo-wide to zero.

+++ internal/search/apply.go
@@ -72,7 +72,7 @@ func (a Applier) Apply(ctx context.Context, tx *sql.Tx, rec statelog.Record, _ s
 
 // embed writes the vector and its sign code.
 func (a Applier) embed(ctx context.Context, tx *sql.Tx, vec VectorRecord, at statelog.Position) (int, error) {
-	version := int64(at.Packed())
+	version := at.Packed()
 	container := vec.Container
 	res, err := tx.ExecContext(ctx, `
 		INSERT INTO kb_vectors
@@ -146,7 +146,7 @@ func (a Applier) embed(ctx context.Context, tx *sql.Tx, vec VectorRecord, at sta
 func (a Applier) forget(ctx context.Context, tx *sql.Tx, subject Subject, at statelog.Position) (int, error) {
 	res, err := tx.ExecContext(ctx,
 		`DELETE FROM kb_vectors WHERE source = ? AND source_id = ? AND version < ?`,
-		string(subject.Source), subject.ID, int64(at.Packed()))
+		string(subject.Source), subject.ID, at.Packed())
 	if err != nil {
 		return 0, fmt.Errorf("search: forget the vector for %s at %s: %w",
 			subject, at, err)
